Refactored Deferred to be a child of PostEffect
