+++ gl_engine/gl_engine.vcxproj
@@ -180,9 +180,9 @@
     <ClInclude Include="node\Node.h" />
     <ClInclude Include="pch.h" />
     <ClInclude Include="post\Bloom.h" />
+    <ClInclude Include="post\Deferred.h" />
     <ClInclude Include="post\PostEffect.h" />
     <ClInclude Include="post\ToneMap.h" />
-    <ClInclude Include="render\DeferredRender.h" />
     <ClInclude Include="render\Renderer.h" />
     <ClInclude Include="render\Window.h" />
     <ClInclude Include="resources\Buffer.h" />
@@ -228,8 +228,8 @@
       <PrecompiledHeader Condition="'$(Configuration)|$(Platform)'=='Release|x64'">Create</PrecompiledHeader>
     </ClCompile>
     <ClCompile Include="post\Bloom.cpp" />
+    <ClCompile Include="post\Deferred.cpp" />
     <ClCompile Include="post\ToneMap.cpp" />
-    <ClCompile Include="render\DeferredRender.cpp" />
     <ClCompile Include="render\Renderer.cpp" />
     <ClCompile Include="render\Window.cpp" />
     <ClCompile Include="resources\Buffer.cpp" />
