+++ gl_engine/x64/Debug/gl_engine.log
@@ -1,4 +1,4 @@
-﻿  DeferredRender.cpp
+﻿  Deferred.cpp
   Renderer.cpp
   Generating Code...
   gl_engine.vcxproj -> E:\DOCS\PROGRAMMING\CPP\OpenGL\repos\opengl\x64\Debug\gl_engine.lib
